Prevent multiple logouts Fixes AB#10697, AB#10711
