+++ Fmas12d.Ui/src/app/app.routes.ts
@@ -1,12 +1,13 @@
 import { AuthorizationGuard } from './authorization.guard';
 import { AutoLoginComponent } from './components/auto-login/auto-login.component';
-import { OnCallDoctorListComponent } 
+import { OnCallDoctorListComponent }
   from './components/on-call-doctor/on-call-doctor-list/on-call-doctor-list.component';
 import { PageNotFoundComponent } from './components/page-not-found/page-not-found.component';
 import { Routes } from '@angular/router';
 import { UnauthorizedComponent } from './components/unauthorized/unauthorized.component';
 import { UserProfileComponent } from './components/user-profile/user-profile.component';
 import { WelcomeComponent } from './components/welcome/welcome.component';
+import { SignOutComponent } from './components/sign-out/signout.component';
 
 export const AppRoutes: Routes = [
   {
@@ -21,6 +22,10 @@ export const AppRoutes: Routes = [
   {
     path: 'unauthorized',
     component: UnauthorizedComponent
+  },
+  {
+    path: 'signout',
+    component: SignOutComponent
   },
   {
     path: 'user/profile',
