+++ Fmas12d.Ui/src/app/components/sign-out/signout.component.html
@@ -1,0 +1,3 @@
+<div class="error-container">
+  <app-delay-spinner [delayMessage]="'Signing Out...'"></app-delay-spinner>
+</div>
