+++ Fmas12d.Ui/src/app/components/sign-out/signout.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-sign-out',
+  templateUrl: './signout.component.html',
+  styleUrls: ['./signout.component.css']
+})
+export class SignOutComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+}
